Rename Timeslot.js to TimeSlot.js
